Rename 'materials' to 'material'
