+++ core/include/detray/builders/homogeneous_material_generator.hpp
@@ -12,9 +12,9 @@
 #include "detray/definitions/algebra.hpp"
 #include "detray/definitions/detail/qualifiers.hpp"
 #include "detray/definitions/indexing.hpp"
-#include "detray/materials/material.hpp"
-#include "detray/materials/material_rod.hpp"
-#include "detray/materials/predefined_materials.hpp"
+#include "detray/material/material.hpp"
+#include "detray/material/material_rod.hpp"
+#include "detray/material/predefined_materials.hpp"
 #include "detray/utils/logging.hpp"
 #include "detray/utils/ranges.hpp"
 #include "detray/utils/type_registry.hpp"
