+++ core/include/detray/builders/material_map_generator.hpp
@@ -15,10 +15,10 @@
 #include "detray/geometry/detail/surface_kernels.hpp"
 #include "detray/geometry/shapes/cylinder2D.hpp"
 #include "detray/geometry/shapes/ring2D.hpp"
-#include "detray/materials/material.hpp"
-#include "detray/materials/material_rod.hpp"
-#include "detray/materials/material_slab.hpp"
-#include "detray/materials/predefined_materials.hpp"
+#include "detray/material/material.hpp"
+#include "detray/material/material_rod.hpp"
+#include "detray/material/material_slab.hpp"
+#include "detray/material/predefined_materials.hpp"
 #include "detray/utils/grid/detail/axis.hpp"
 #include "detray/utils/logging.hpp"
 #include "detray/utils/ranges.hpp"
