+++ core/include/detray/geometry/detail/volume_kernels.hpp
@@ -11,9 +11,9 @@
 #include "detray/definitions/algebra.hpp"
 #include "detray/definitions/detail/qualifiers.hpp"
 #include "detray/definitions/indexing.hpp"
-#include "detray/materials/concepts.hpp"
-#include "detray/materials/detail/material_accessor.hpp"
-#include "detray/materials/material.hpp"
+#include "detray/material/concepts.hpp"
+#include "detray/material/detail/material_accessor.hpp"
+#include "detray/material/material.hpp"
 #include "detray/navigation/accelerators/search_window.hpp"
 #include "detray/navigation/concepts.hpp"
