+++ core/include/detray/material/detail/density_effect_data.hpp
