+++ core/include/detray/propagator/actors/pointwise_material_interactor.hpp
@@ -12,9 +12,9 @@
 #include "detray/definitions/algebra.hpp"
 #include "detray/definitions/detail/qualifiers.hpp"
 #include "detray/definitions/track_parametrization.hpp"
-#include "detray/materials/concepts.hpp"
-#include "detray/materials/detail/material_accessor.hpp"
-#include "detray/materials/interaction.hpp"
+#include "detray/material/concepts.hpp"
+#include "detray/material/detail/material_accessor.hpp"
+#include "detray/material/interaction.hpp"
 #include "detray/propagator/actors/parameter_updater.hpp"
 #include "detray/propagator/base_actor.hpp"
 #include "detray/tracks/bound_track_parameters.hpp"
