+++ core/include/detray/propagator/rk_stepper.hpp
@@ -12,7 +12,8 @@
 #include "detray/definitions/algebra.hpp"
 #include "detray/definitions/detail/qualifiers.hpp"
 #include "detray/definitions/units.hpp"
-#include "detray/materials/interaction.hpp"
+#include "detray/material/interaction.hpp"
+#include "detray/material/material.hpp"
 #include "detray/navigation/policies.hpp"
 #include "detray/propagator/base_stepper.hpp"
 #include "detray/propagator/transport_jacobian.hpp"
