+++ core/include/detray/utils/consistency_checker.hpp
@@ -10,8 +10,8 @@
 // Project include(s)
 #include "detray/geometry/surface.hpp"
 #include "detray/geometry/tracking_volume.hpp"
-#include "detray/materials/concepts.hpp"
-#include "detray/materials/predefined_materials.hpp"
+#include "detray/material/concepts.hpp"
+#include "detray/material/predefined_materials.hpp"
 #include "detray/utils/logging.hpp"
 #include "detray/utils/ranges.hpp"
 #include "detray/utils/type_registry.hpp"
