+++ detectors/include/detray/detectors/default_metadata.hpp
@@ -18,9 +18,9 @@
 #include "detray/geometry/shapes/unbounded.hpp"
 #include "detray/geometry/shapes/unmasked.hpp"
 #include "detray/geometry/surface_descriptor.hpp"
-#include "detray/materials/material_map.hpp"
-#include "detray/materials/material_rod.hpp"
-#include "detray/materials/material_slab.hpp"
+#include "detray/material/material_map.hpp"
+#include "detray/material/material_rod.hpp"
+#include "detray/material/material_slab.hpp"
 #include "detray/navigation/accelerators/brute_force.hpp"
 #include "detray/navigation/accelerators/spatial_grid.hpp"
