+++ io/include/detray/io/backend/material_map_reader.hpp
@@ -15,7 +15,7 @@
 #include "detray/io/backend/detail/type_info.hpp"
 #include "detray/io/backend/homogeneous_material_reader.hpp"
 #include "detray/io/frontend/payloads.hpp"
-#include "detray/materials/material_slab.hpp"
+#include "detray/material/material_slab.hpp"
 #include "detray/utils/type_list.hpp"
 
 // System include(s)
