+++ io/include/detray/io/backend/material_map_writer.hpp
@@ -12,7 +12,7 @@
 #include "detray/io/backend/detail/type_info.hpp"
 #include "detray/io/backend/homogeneous_material_writer.hpp"
 #include "detray/io/frontend/payloads.hpp"
-#include "detray/materials/material_slab.hpp"
+#include "detray/material/material_slab.hpp"
 
 // System include(s)
 #include <string_view>
