+++ tests/include/detray/test/common/build_toy_detector.hpp
@@ -23,8 +23,8 @@
 #include "detray/definitions/units.hpp"
 #include "detray/detectors/toy_metadata.hpp"
 #include "detray/geometry/tracking_volume.hpp"
-#include "detray/materials/mixture.hpp"
-#include "detray/materials/predefined_materials.hpp"
+#include "detray/material/mixture.hpp"
+#include "detray/material/predefined_materials.hpp"
 #include "detray/utils/consistency_checker.hpp"
 #include "detray/utils/print_detector.hpp"
 #include "detray/utils/ranges.hpp"
