+++ tests/include/detray/test/common/factories/endcap_generator.hpp
@@ -16,7 +16,7 @@
 #include "detray/definitions/indexing.hpp"
 #include "detray/geometry/mask.hpp"
 #include "detray/geometry/shapes/rectangle2D.hpp"
-#include "detray/materials/material.hpp"
+#include "detray/material/material.hpp"
 #include "detray/utils/logging.hpp"
 
 // System include(s)
