+++ tests/integration_tests/cpu/material/material_interaction.cpp
@@ -10,10 +10,10 @@
 #include "detray/definitions/units.hpp"
 #include "detray/geometry/mask.hpp"
 #include "detray/geometry/shapes/unbounded.hpp"
-#include "detray/materials/interaction.hpp"
-#include "detray/materials/material.hpp"
-#include "detray/materials/material_slab.hpp"
-#include "detray/materials/predefined_materials.hpp"
+#include "detray/material/interaction.hpp"
+#include "detray/material/material.hpp"
+#include "detray/material/material_slab.hpp"
+#include "detray/material/predefined_materials.hpp"
 #include "detray/navigation/caching_navigator.hpp"
 #include "detray/propagator/actors.hpp"
 #include "detray/propagator/actors/parameter_updater.hpp"
