+++ tests/unit_tests/cpu/core/detector.cpp
@@ -9,7 +9,7 @@
 #include "detray/core/detector.hpp"
 
 #include "detray/definitions/indexing.hpp"
-#include "detray/materials/predefined_materials.hpp"
+#include "detray/material/predefined_materials.hpp"
 
 // Detray test include(s)
 #include "detray/test/framework/types.hpp"
